REFECTOR: Chart.js v4 treeshaking 으로변경 -> 번들링크기 감소
