+++ package.json
@@ -57,7 +57,6 @@
     ]
   },
   "devDependencies": {
-    "@types/axios": "^0.14.0",
     "@types/chart.js": "^2.9.38",
     "@types/react-router-dom": "^5.3.3",
     "@types/react-tooltip": "^4.2.4",
