+++ src/components/Graph.tsx
@@ -1,9 +1,41 @@
-import 'chart.js/auto';
-import { Chart, getElementsAtEvent } from 'react-chartjs-2';
+import { Chart as ChartComponent, getElementsAtEvent } from 'react-chartjs-2';
+// import 'chart.js/auto';
 import { TimeSeriesData } from 'types';
 import 'chartjs-adapter-moment';
-import { ChartData, ChartOptions, ChartTypeRegistry } from 'chart.js';
 import { useRef } from 'react';
+
+import {
+  Title,
+  Filler,
+  BarController,
+  ChartData,
+  ChartOptions,
+  ChartTypeRegistry,
+  Chart as ChartJS,
+  LineElement,
+  PointElement,
+  LinearScale,
+  CategoryScale,
+  BarElement,
+  Legend,
+  Tooltip,
+  TimeScale,
+} from 'chart.js';
+
+// Register your components
+ChartJS.register(
+  Title,
+  Filler,
+  BarController,
+  TimeScale,
+  LinearScale,
+  CategoryScale,
+  BarElement,
+  PointElement,
+  LineElement,
+  Legend,
+  Tooltip
+);
 
 interface Props {
   data: TimeSeriesData;
@@ -136,7 +168,7 @@ export default function Graph({ data, selectedId, handleFilterById }: Props) {
   };
 
   return (
-    <Chart
+    <ChartComponent
       type="bar"
       data={config}
       options={options}
